requirement.txt
annotated-types==0.7.0
anyio==4.3.0
astor==0.8.1
blinker==1.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dbus-python==1.2.16
decorator==5.1.1
distro==1.9.0
distro-info==0.23+ubuntu1.1
exceptiongroup==1.2.0
filelock==3.13.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==2.8
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.8.2
jsonify==0.5
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnx==1.16.1
onnxruntime-gpu==1.17.1
openai==1.59.8
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
paddlepaddle-gpu==2.6.1
pandas==2.2.2
Pillow==9.5.0
protobuf==5.26.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pydantic==2.10.5
pydantic_core==2.27.2
PyGObject==3.36.0
PyMuPDF==1.26.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
reportlab==4.4.2
requests==2.32.3
requests-unixsocket==0.2.0
scikit-image==0.22.0
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.3
six==1.14.0
sniffio==1.3.1
sympy==1.12
tifffile==2024.2.12
torch @ file:///home/to/torch-1.10.2%2Bcu111-cp39-cp39-linux_x86_64.whl#sha256=78ae2260ac1902c5a224fe1b7994072bbfa7ef8a1734d8697ca74249fa6391c5
torchaudio==0.12.1+cu116
torchvision @ file:///home/to/torchvision-0.11.3%2Bcu111-cp39-cp39-linux_x86_64.whl#sha256=0644e5b680e4f506d92a3e8ad67b34e18d18c1e10a4a8ffc662cfa909fd70ecf
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.48
ultralytics-thop==2.0.0
unattended-upgrades==0.1
Unidecode==1.3.8
urllib3==1.25.8
Werkzeug==3.0.2
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
